pyproject.toml
[project]
name = "nind-denoise"
version = "0.3.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asttokens~=3.0.0",
    "beautifulsoup4~=4.13.4",
    "bs4~=0.0.2",
    "configargparse~=1.7.1",
    "decorator~=5.2.1",
    "docopt-ng~=0.9.0",
    "executing~=2.2.0",
    "exiv2~=0.17.3",
    "filelock>=3.18.0",
    "fsspec>=2025.5.1",
    "imagecodecs~=2025.8.2",
    "imageio~=2.37.0",
    "jedi~=0.19.2",
    "jinja2~=3.1.4",
    "lxml~=6.0.0",
    "markupsafe>=3.0.2",
    "matplotlib-inline~=0.1.7",
    "mpmath~=1.3.0",
    "networkx~=3.3",
    "numpy~=2.1.2",
    "opencv-contrib-python~=4.12.0.88",
    "opencv-python~=4.12.0.88",
    "ostools~=1.1.8",
    "parso~=0.8.4",
    "path~=17.1.0",
    "pexpect~=4.9.0",
    "pillow>=11.3.0",
    "pip>=25.2",
    "piqa~=1.3.2",
    "prompt-toolkit~=3.0.51",
    "ptyprocess~=0.7.0",
    "pure-eval~=0.2.3",
    "pygments~=2.19.2",
    "pyyaml~=6.0.2",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "soupsieve~=2.7",
    "stack-data~=0.6.3",
    "sympy>=1.14.0",
    "tifffile~=2025.6.11",
    "torch~=2.8.0",
    "torchaudio~=2.8.0",
    "torchvision~=0.23.0",
    "traitlets~=5.14.3",
    "typing-extensions>=4.14.1",
    "wcwidth~=0.2.13",
]
[dependency-groups]
dev = [
    "ipython>=9.5.0"
]
test = [
    "nind-denoise",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[tool.uv.sources]
nind-denoise = { workspace = true }
